+++ bindings/core/src/method_handler/account.rs
@@ -6,7 +6,8 @@ use std::str::FromStr;
 
 use iota_sdk::{
     client::api::{
-        PreparedTransactionData, PreparedTransactionDataDto, SignedTransactionData, SignedTransactionDataDto,
+        input_selection::Burn, PreparedTransactionData, PreparedTransactionDataDto, SignedTransactionData,
+        SignedTransactionDataDto,
     },
     types::block::{
         output::{dto::OutputDto, Output, Rent},
@@ -26,24 +27,10 @@ use crate::{method::AccountMethod, Response, Result};
 
 pub(crate) async fn call_account_method_internal(account: &Account, method: AccountMethod) -> Result<Response> {
     let response = match method {
-        AccountMethod::BurnNativeToken {
-            token_id,
-            burn_amount,
-            options,
-        } => {
+        AccountMethod::Burn { burn, options } => {
             let transaction = account
-                .burn_native_token(
-                    token_id,
-                    U256::try_from(&burn_amount).map_err(|_| Error::InvalidField("burn_amount"))?,
-                    options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
-                )
-                .await?;
-            Response::SentTransaction(TransactionDto::from(&transaction))
-        }
-        AccountMethod::BurnNft { nft_id, options } => {
-            let transaction = account
-                .burn_nft(
-                    nft_id,
+                .burn(
+                    Burn::try_from(&burn)?,
                     options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                 )
                 .await?;
@@ -71,24 +58,6 @@ pub(crate) async fn call_account_method_internal(account: &Account, method: Acco
                 .await?;
             Response::SentTransaction(TransactionDto::from(&transaction))
         }
-        AccountMethod::DestroyAlias { alias_id, options } => {
-            let transaction = account
-                .destroy_alias(
-                    alias_id,
-                    options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
-                )
-                .await?;
-            Response::SentTransaction(TransactionDto::from(&transaction))
-        }
-        AccountMethod::DestroyFoundry { foundry_id, options } => {
-            let transaction = account
-                .destroy_foundry(
-                    foundry_id,
-                    options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
-                )
-                .await?;
-            Response::SentTransaction(TransactionDto::from(&transaction))
-        }
         AccountMethod::GenerateAddresses { amount, options } => {
             let address = account.generate_addresses(amount, options).await?;
             Response::GeneratedAddress(address)
@@ -111,17 +80,12 @@ pub(crate) async fn call_account_method_internal(account: &Account, method: Acco
             let transaction = account.get_transaction(&transaction_id).await;
             Response::Transaction(transaction.as_ref().map(TransactionDto::from).map(Box::new))
         }
-        AccountMethod::GetIncomingTransactionData { transaction_id } => {
-            let transaction = account.get_incoming_transaction_data(&transaction_id).await;
+        AccountMethod::GetIncomingTransaction { transaction_id } => {
+            let transaction = account.get_incoming_transaction(&transaction_id).await;
 
             transaction.map_or_else(
-                || Response::IncomingTransactionData(None),
-                |transaction| {
-                    Response::IncomingTransactionData(Some(Box::new((
-                        transaction_id,
-                        TransactionDto::from(&transaction),
-                    ))))
-                },
+                || Response::Transaction(None),
+                |transaction| Response::Transaction(Some(Box::new(TransactionDto::from(&transaction)))),
             )
         }
         AccountMethod::Addresses => {
@@ -141,20 +105,15 @@ pub(crate) async fn call_account_method_internal(account: &Account, method: Acco
             Response::OutputsData(outputs.iter().map(OutputDataDto::from).collect())
         }
         AccountMethod::IncomingTransactions => {
-            let transactions = account.incoming_transactions().await?;
-            Response::IncomingTransactionsData(
-                transactions
-                    .into_iter()
-                    .map(|d| (d.0, TransactionDto::from(&d.1)))
-                    .collect(),
-            )
+            let transactions = account.incoming_transactions().await;
+            Response::Transactions(transactions.iter().map(TransactionDto::from).collect())
         }
         AccountMethod::Transactions => {
-            let transactions = account.transactions().await?;
+            let transactions = account.transactions().await;
             Response::Transactions(transactions.iter().map(TransactionDto::from).collect())
         }
         AccountMethod::PendingTransactions => {
-            let transactions = account.pending_transactions().await?;
+            let transactions = account.pending_transactions().await;
             Response::Transactions(transactions.iter().map(TransactionDto::from).collect())
         }
         AccountMethod::DecreaseNativeTokenSupply {