+++ sdk/examples/how_tos/accounts_and_addresses/create_address.rs
@@ -1,0 +1,31 @@
+// Copyright 2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+//! In this example we will generate an address.
+//! Rename `.env.example` to `.env` first.
+//!
+//! `cargo run --release --all-features --example create_address`
+
+use iota_sdk::wallet::{Result, Wallet};
+
+#[tokio::main]
+async fn main() -> Result<()> {
+    // This example uses secrets in environment variables for simplicity which should not be done in production.
+    dotenvy::dotenv().ok();
+
+    // Create the wallet
+    let wallet = Wallet::builder().finish().await?;
+
+    // Get the account we generated with `01_create_wallet`
+    let account = wallet.get_account("Alice").await?;
+
+    // Set the stronghold password
+    wallet
+        .set_stronghold_password(&std::env::var("STRONGHOLD_PASSWORD").unwrap())
+        .await?;
+
+    let address = account.generate_addresses(1, None).await?;
+    println!("Generated address: {}", address[0].address());
+
+    Ok(())
+}