+++ sdk/src/wallet/message_interface/message_handler.rs
@@ -25,7 +25,7 @@ use crate::{
     types::block::{
         output::{
             dto::{OutputBuilderAmountDto, OutputDto},
-            AliasOutput, BasicOutput, FoundryOutput, NftOutput, Output, Rent,
+            AliasOutput, BasicOutput, FoundryOutput, NativeToken, NftOutput, Output, Rent,
         },
         Error,
     },
@@ -497,9 +497,8 @@ impl WalletMessageHandler {
             } => {
                 convert_async_panics(|| async {
                     let transaction = account
-                        .burn_native_token(
-                            token_id,
-                            U256::try_from(&burn_amount).map_err(|_| Error::InvalidField("burn_amount"))?,
+                        .burn(
+                            NativeToken::new(token_id, burn_amount)?,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
@@ -510,7 +509,7 @@ impl WalletMessageHandler {
             AccountMethod::BurnNft { nft_id, options } => {
                 convert_async_panics(|| async {
                     let transaction = account
-                        .burn_nft(
+                        .burn(
                             nft_id,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
@@ -550,7 +549,7 @@ impl WalletMessageHandler {
             AccountMethod::DestroyAlias { alias_id, options } => {
                 convert_async_panics(|| async {
                     let transaction = account
-                        .destroy_alias(
+                        .burn(
                             alias_id,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
@@ -562,7 +561,7 @@ impl WalletMessageHandler {
             AccountMethod::DestroyFoundry { foundry_id, options } => {
                 convert_async_panics(|| async {
                     let transaction = account
-                        .destroy_foundry(
+                        .burn(
                             foundry_id,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
@@ -597,15 +596,14 @@ impl WalletMessageHandler {
                     transaction.as_ref().map(TransactionDto::from).map(Box::new),
                 ))
             }
-            AccountMethod::GetIncomingTransactionData { transaction_id } => {
-                let transaction = account.get_incoming_transaction_data(&transaction_id).await;
+            AccountMethod::GetIncomingTransaction { transaction_id } => {
+                let transaction = account.get_incoming_transaction(&transaction_id).await;
 
                 transaction.map_or_else(
-                    || Ok(Response::IncomingTransactionData(None)),
+                    || Ok(Response::Transaction(None)),
                     |transaction| {
-                        Ok(Response::IncomingTransactionData(Some(Box::new((
-                            transaction_id,
-                            TransactionDto::from(&transaction),
+                        Ok(Response::Transaction(Some(Box::new(TransactionDto::from(
+                            &transaction,
                         )))))
                     },
                 )
@@ -627,22 +625,19 @@ impl WalletMessageHandler {
                 Ok(Response::OutputsData(outputs.iter().map(OutputDataDto::from).collect()))
             }
             AccountMethod::IncomingTransactions => {
-                let transactions = account.incoming_transactions().await?;
-                Ok(Response::IncomingTransactionsData(
-                    transactions
-                        .into_iter()
-                        .map(|d| (d.0, TransactionDto::from(&d.1)))
-                        .collect(),
+                let transactions = account.incoming_transactions().await;
+                Ok(Response::Transactions(
+                    transactions.iter().map(TransactionDto::from).collect(),
                 ))
             }
             AccountMethod::Transactions => {
-                let transactions = account.transactions().await?;
+                let transactions = account.transactions().await;
                 Ok(Response::Transactions(
                     transactions.iter().map(TransactionDto::from).collect(),
                 ))
             }
             AccountMethod::PendingTransactions => {
-                let transactions = account.pending_transactions().await?;
+                let transactions = account.pending_transactions().await;
                 Ok(Response::Transactions(
                     transactions.iter().map(TransactionDto::from).collect(),
                 ))