+++ sdk/CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ignore_if_bech32_mismatch` parameter to `Wallet::restore_backup()`;
 - `OutputWithMetadata::{into_output, into_metadata}` methods;
 - Storage and Backup migration;
+- `types::block::Error::InvalidFoundryZeroSerialNumber` variant;
 
 ### Changed
 
@@ -55,6 +56,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `InputSigningDataDto::chain` is now `Vec<u32>` instead of `Chain`;
 - Most `StrongholdAdapter` fns no longer require a mutable reference;
 - `StorageProvider` fns no longer require a mutable reference;
+- `Account::burn_native_tokens()`, `Account::burn_nft()`, `Account::destroy_foundry()`, `Account::destroy_alias()` merged into `Account::burn()`;
+- `Account::burn_native_tokens()`, `Account::burn_nft()`, `Account::destroy_foundry()`, `Account::destroy_alias()` merged into `Account::burn()`
+- `ValidationContext::{input_native_tokens, output_native_tokens}` from HashMap to BTreeMap;
+- Rename `AccountInner::get_incoming_transaction_data` to `get_incoming_transaction`;
+- `AccountInner::{incoming_transactions, transactions, pending_transactions}` don't return a `Result` anymore;
+- `AccountInner::incoming_transactions` returns a `Vec` instead of a `HashMap`;
+- `{MetadataFeature, TagFeature}::new` take an `impl Into<Vec<u8>>` param;
 
 ### Removed
 
@@ -69,6 +77,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Storage records decryption;
+
+### Fixed
+
+- Validation for transitions in the input selection;
+- Automatically increase foundry counter of alias outputs;
+- Validate that foundry outputs can't have serial number `0`;
 
 ## 0.3.0 - 2023-05-02